feat: Add quotation number padding setting
Introduces a new `quotationPadding` setting to control the zero-padding of quotation numbers. This allows for more flexibility in generating sequential quotation identifiers, improving consistency and professionalism.

The default padding is set to 6, ensuring that numbers like "1" become "000001". The UI in the Settings page has been updated to accommodate this new option.